LayoutTests/platform/ios-simulator/editing/selection/context-menu-on-text-expected.txt

CONSOLE MESSAGE: line 24: TypeError: undefined is not an object (evaluating 'items.length')
This test checks that the context menu is shown over a simple non-editable text.

Lorem ipsum
FAIL
